services/backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.4.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0
greenlet>=3.0.0

# Redis and Celery
redis>=5.0.0
celery>=5.3.0

# WebSocket and Real-time Communication
python-socketio>=5.10.0
aioredis>=2.0.0
python-socketio[asyncio_client]>=5.10.0

# Time series and monitoring
influxdb-client>=1.38.0

# MQTT
paho-mqtt>=1.6.0

# Object Storage
minio>=7.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt,argon2]>=1.7.0
python-multipart>=0.0.6
argon2-cffi>=23.1.0

# HTTP client
httpx>=0.25.0
aiohttp>=3.9.0

# Logging
python-json-logger>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-xdist>=3.3.0
pytest-clarity>=1.0.0
factory-boy>=3.3.0
faker>=20.0.0
httpx>=0.25.0

# CLI tools
click>=8.1.0

# Testing dependencies
aiosqlite>=0.21.0